Develop

+++ base/views.py
@@ -1,10 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate,login,logout
 from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models import Q
-from .models import Room, Topic
+from .models import Room, Topic,Message
 from .forms import RoomForm
 
 # Create your views here.
@@ -14,8 +16,11 @@ from .forms import RoomForm
 #    {'id':3,'name':'Frontend Developers!'}
 #]
 def loginpage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method=='POST':
-        username=request.POST.get('username')
+        username=request.POST.get('username').lower()
         password=request.POST.get('password')
         try:
             user=User.objects.get(username=username)
@@ -27,7 +32,22 @@ def loginpage(request):
             return redirect('home')
         else:
             messages.error(request, "Username or Password does not exist.")
-    context={}
+    context={'page':page}
+    return render(request,'base/login_register.html',context)
+
+def registerpage(request):
+    form=UserCreationForm()
+    if request.method=='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during the registration')
+    context={'form':form}
     return render(request,'base/login_register.html',context)
 
 def home(request):
@@ -42,7 +62,15 @@ def home(request):
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    context={'room':room}
+    room_messages=room.message_set.all().order_by('-created')
+    if request.method=='POST':
+        message=Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        return redirect('room',pk=room.id)
+    context = {'room':room,'room_messages':room_messages}
     return render(request, 'base/room.html',context)
 
 @login_required(login_url='/login')
@@ -56,9 +84,12 @@ def createroom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def updateroom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     if request.method=='POST':
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -67,8 +98,11 @@ def updateroom(request,pk):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def deleteroom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     if request.method=='POST':
         room.delete()
         return redirect('home')
